AMS-76: move decorator for view type switcher
